+++ Sources/csuper-gui/main_window.h
@@ -55,5 +55,9 @@ G_MODULE_EXPORT gboolean setButtonMainWindowClipboardSensitive(gpointer data);
 void readMainWindowSize(globalData *data);
 G_MODULE_EXPORT gboolean saveMainWindowSize(GtkWidget *widget,GdkEvent *event,gpointer user_data);
 G_MODULE_EXPORT void changeDistributorButton(GtkWidget *widget, gpointer data);
+G_MODULE_EXPORT void updateCalculatorMainWindow(GtkWidget *widget, gpointer data);
+G_MODULE_EXPORT void validateCalculatorPoints(GtkWidget *widget, gpointer data);
+void fillCalculatorNames(globalData *data);
+void updateMainWindowSide(globalData *data);
 
 #endif // MAIN_WINDOW_H_INCLUDED